requirements.txt
torch
unsloth
Pillow
requests
transformers
trl
torchvision
